+++ src/components/QueryBuilder/QueryBuilder.js
@@ -42,6 +42,7 @@ function QueryBuilder({ ldp, initialState, stateHasChanged, metadataHasChanged, 
   const [queryResponse, setQueryResponse] = useState({ key: null, resp: [] });
   const [showNewModal, setShowNewModal] = useState(false);
   const [showSaveModal, setShowSaveModal] = useState(false);
+  const [showCopyModal, setShowCopyModal] = useState(false);
   const [alreadyExecuted, setAlreadyExecuted] = useState(false);
   const showDevInfo = stripes.config?.showDevInfo;
   const onSubmit = values => loadResults(intl, stripes, values, setQueryResponse, setError);
@@ -56,6 +57,21 @@ function QueryBuilder({ ldp, initialState, stateHasChanged, metadataHasChanged, 
   const newQuery = async () => {
     onClear();
     setShowNewModal(false);
+  };
+
+  const copyQuery = async () => {
+    const newState = {
+      ...initialState,
+      META: { ...initialState.META },
+    };
+    delete newState.META.id;
+    newState.META.displayName = intl.formatMessage(
+      { id: 'ui-ldp.copy-of' },
+      { string: newState.META.displayName }
+    );
+
+    metadataHasChanged(newState);
+    setShowCopyModal(false);
   };
 
   return (
@@ -112,6 +128,12 @@ function QueryBuilder({ ldp, initialState, stateHasChanged, metadataHasChanged, 
                                 onClick={() => setShowSaveModal(true)}
                                 data-cy={`${namePrefix}.saveQuery`}
                               />
+                              <IconButton
+                                icon="duplicate"
+                                aria-label={intl.formatMessage({ id: 'ui-ldp.button.copy-query' })}
+                                onClick={() => setShowCopyModal(true)}
+                                data-cy={`${namePrefix}.copyQuery`}
+                              />
                             </>
                           }
                         >
@@ -161,6 +183,14 @@ function QueryBuilder({ ldp, initialState, stateHasChanged, metadataHasChanged, 
                     metadataHasChanged={metadataHasChanged}
                   />
               }
+              <ConfirmationModal
+                open={showCopyModal}
+                heading={<FormattedMessage id="ui-ldp.button.copy-query" />}
+                message={<FormattedMessage id="ui-ldp.desc.copy-query" />}
+                confirmLabel={<FormattedMessage id="ui-ldp.button.copy-query" />}
+                onConfirm={copyQuery}
+                onCancel={() => setShowCopyModal(false)}
+              />
             </form>
           );
         }}