+++ src/components/QueryBuilder/SaveQueryModal.js
@@ -43,7 +43,7 @@ function SaveQueryModal({ onClose, queryFormValues, autoUpdateName, metadataHasC
     } else {
       method = 'POST';
       path = '/settings/entries';
-      content.META.id = id = uuidv4();
+      content.META.id = id = uuidv4(); // eslint-disable-line no-multi-assign
     }
 
     const res = await stripesFetch(stripes, path, {
@@ -51,7 +51,7 @@ function SaveQueryModal({ onClose, queryFormValues, autoUpdateName, metadataHasC
       body: JSON.stringify({
         id,
         scope: 'ui-ldp.queries',
-        key: values.name,
+        key: id, // We don't actually use this, it's just a disambigutor
         value: content,
       }),
     });